Introduce a new hint error to display at the end of a failed execution (#3519)
In a recent previous PR the existing hint.Error was refactored in order
to allow for hints to be displayed at the end of a failed execution in a
more consistent way compared to other help messages.

In this PR that change was slightly reverted so that it would be easier
to remove the deprecated hint.Error
(`WrapWithDisplay` is reverted back to `Wrap`)

So instead, this PR introduced a new hint.Error, which if used, will
show the hint at the end of an execution like other help messages. In
addition the mechanism was improved such that even hints going through
gRPC will be properly identified.

Finally, this new hint error is used in cases where a cloud secret could
not be found.