Wildcard BUILDs (#3661)
This PR is my first pass on wildcard BUILD expansion. That is, builds in
the form of `BUILD ./*+target`. This code will attempt to expand the
BUILD into multiple commands if both the Earthfile & the target exist.
The code will only explore directories 1 level beneath the referenced
directory and globstar (`**`) is not yet supported.

Example: 

```
$ tree tests/wildcard-build 
tests/wildcard-build
├── bar
│   └── Earthfile
├── baz
│   └── Earthfile
├── foo
│   └── Earthfile
├── no-earthfile
└── no-target
    └── Earthfile
```

```
VERSION 0.7

test:
    FROM alpine
    BUILD ./wildcard-build/*+test
```

https://cloud.earthly.dev/builds/ec88fff1-d353-4045-a345-e3ce10cadff0


![image](https://github.com/earthly/earthly/assets/332408/be80f980-9b96-4d1b-9a90-12ee90d897ba)