Fix worklow that merges main to docs-0.8 (#3893)
Before this change, the target `+merge-main-to-docs` would open a new
temp PR and wait for all its GH checks to complete (or for first
failure). Unfortunately, One of those checks that it was waiting for is
the job that invoked the target, so the target would get stuck waiting
on itself until timing out.

This change limits the target to only watch for the "check broken links"
job to complete before trying to merge the PR/branch.