Run `--raw-output`  (#4023)
Re https://github.com/earthly/earthly/issues/3713

There are two ways to solve this, one is adding RawOutput bool to
DeltaLog and wiring it through. That would change the logstream api.

This is the other solution, which is simple but hacky. We just check for
`--raw-output` on the command before printing. @alexcb suggested
starting here.