remove invalid test
This test was passing `should-fail=true` rather than `should_fail=true`;
however it's not clear what this test was doing as earthly was passing
so changing the argument to `should_fail=true` then causes this test to
fail (since earthly was not failing).

There was also a `test-fails` target, which was never used anywhere.
Ironically this target was also passing, so it wasn't a matter of simply
changing

    DO --pass-args +RUN_EARTHLY_ARGS --target=/the-test+test --should-fail=true

to

    DO --pass-args +RUN_EARTHLY_ARGS --target=/the-test+test-fails --should_fail=true

Signed-off-by: Alex Couture-Beil <alex@earthly.dev>