Disable cloud upsell at the end of build
This is implemented as a separte flag instead of piggy-backing on
EARTHLY_DISABLE_ANALYTICS as suggested in #3601 because the purpose of
these two things seem separate.

urfave/cli doesn't support env-only flags. The EARTHLY_DISABLE_ANALYTICS
flag was simply defined with no name, which makes it show up in --help
but seemingly without a --option. However, defining two flags that way
doesn't work, as now we have 2 flags named empty string and the flagset
will panic.

So instead I've opted for a hidden flag. Unfortunately this causes no
documentation related to this environment variable to be present in
--help.

Fixes: #3601