Fix device name parsing for MTU discovery when using link scoped devices (#4233)
This fix is relevant for tun devices, which are common in VPN
implementations. For example, while a "normal" default route may look
like:

default via 192.168.1.1 dev enp1s0 proto dhcp src 192.168.1.123 metric
100

A tun device for a VPN may instead look like:

	default dev tun0 scope link

In the current entrypoint, the device name would be incorrectly parsed
as "link" instead of "tun0". By using sed to find the word after "dev",
the result is more likely to be correct. Sed is used elsewhere in the
entrypoint, so it is already an expected utility.

When the device name is incorrectly parsed, the earthly-buildkitd
container fails with the error:

	cat: can't open '/sys/class/net/link/mtu': No such file or directory