Remove old docker objects with --docker-cache
When using a cached data-root, it is possible that some undesired docker
objects have been left behind from a previous execution. In these cases,
we should remove the unnecessary docker objects prior to executing the
RUN.

This change assumes that the intention of the --docker-cache feature is
only to retain image layers between executions, and that it is not
intended to retain volumes, networks, or containers.