Dockerd wrapper housekeeping (#4267) (#4268)
When using a self-hosted satellite with the `--docker-cache`
experimental feature enabled, it was found that leftover bad state in
the cached data-root could cause builds to start failing. In particular,
a leftover, stopped container making use of one of the images that the
`dockerd-wrapper.sh` attempts to `rmi` would result in a `unable to
remove repository reference` error. Some additional debugging revealed
that other types of docker objects are potentially left behind in the
data-root. This change modifies the `docker-wrapper.sh` to do some
housekeeping of these undesirable objects, and also makes changes to
avoid having builds fail due to unanticipated leftover state.

Co-authored-by: eliottwiener-gridunity <134441321+eliottwiener-gridunity@users.noreply.github.com>